Removed all Ascension III/demo assets, added substitutes for third-party dependencies, cleaned up folders and credits, updated core to sync with private branch
Squashed commit of the following:

commit 590bc52d7fdcf56996744f8c785a72552c617cc9
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 16:49:05 2019 -0700

    rewording

commit 39994113fe400489496ff7a57d670fe53cdc5dd9
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 16:47:53 2019 -0700

    removed unused things, revised documentation

commit d3708f5873793bfc060a4bf6a08a53488d6de4e3
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 16:12:20 2019 -0700

    removed some gitignores, moved some fonts

commit a7e4706161213af54abed075f68820372ea71d41
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 16:08:17 2019 -0700

    new titlepic/loadingpic

commit 2ab54b04c5e5cc1f06f43e5e5b985566e24b2376
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 15:58:56 2019 -0700

    removed a bunch of unused stuff

commit c80917574ba256cd1d9c2b0674081bae3b58a0d9
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 15:54:27 2019 -0700

    added substitute rifle

commit d882a36e2235343e175b6dc3e9f0f4e926dbd732
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 15:29:32 2019 -0700

    removed a bunch of audio, removed devconsole

commit 2d4455775eb7772b0c8faea6a24bf578757022e3
Author: XCVGVCX <xcvgvcx@gmail.com>
Date:   Mon Sep 23 15:10:06 2019 -0700

    updated and removed a few things

+++ Assets/CommonCore/CoreShared/State/GameStateBase.cs
@@ -145,14 +145,14 @@ namespace CommonCore.State
         public bool InitialLoaded { get; private set; }
 
         [JsonProperty]
-        private int CurrentUID;
+        private long CurrentUID;
 
         /// <summary>
         /// The next available unique ID
         /// </summary>
         /// <remarks>Accessing this will increment the backing counter (<see cref="CurrentUID"/>)</remarks>
         [JsonIgnore]
-        public int NextUID { get { return ++CurrentUID; } }
+        public long NextUID { get { return ++CurrentUID; } }
 
         /// <summary>
         /// Decorate methods with this atrribute to have them run on GameState initialization. Higher priority is sooner.
